Keep wearable consent within PWA budgets
