+++ service-worker.js
@@ -540,16 +540,26 @@ const APP_SHELL = [ // Includes dynamic chat and Knowledge Base modules for firs
   '/js/touch-tooltip-runtime.js',
   '/js/touch-tooltip.js',
   '/js/url-safety.js',
-  '/js/wearable-adapters.js', '/js/wearables.js',
-  '/js/wearables-strip-actions.js', '/js/wearables-bp-detail-chart.js',
-  '/js/wearables-detail-runtime.js', '/js/wearables-detail-modal.js',
-  '/js/wearables-manual-detail.js', '/js/wearables-formatters.js',
-  '/js/wearables-manual-form-ui.js', '/js/wearables-runtime.js',
-  '/js/wearables-auth-runtime.js', '/js/wearables-settings-runtime.js',
-  '/js/wearables-settings-groups.js', '/js/wearable-relay-consent.js', '/js/wearables-settings-panel.js',
-  '/js/wearables-store.js', '/js/wearables-summary.js',
-  '/js/wearables-connect-runtime.js', '/js/wearables-connect-loader.js',
-  '/js/wearables-connect.js', '/js/wearables-disconnect-recovery.js',
+  '/js/wearable-adapters.js',
+  '/js/wearables.js',
+  '/js/wearables-strip-actions.js',
+  '/js/wearables-bp-detail-chart.js',
+  '/js/wearables-detail-runtime.js',
+  '/js/wearables-detail-modal.js',
+  '/js/wearables-manual-detail.js',
+  '/js/wearables-formatters.js',
+  '/js/wearables-manual-form-ui.js',
+  '/js/wearables-runtime.js',
+  '/js/wearables-auth-runtime.js',
+  '/js/wearables-settings-runtime.js',
+  '/js/wearables-settings-groups.js',
+  '/js/wearables-settings-panel.js',
+  '/js/wearables-store.js',
+  '/js/wearables-summary.js',
+  '/js/wearables-connect-runtime.js',
+  '/js/wearables-connect-loader.js',
+  '/js/wearables-connect.js',
+  '/js/wearables-disconnect-recovery.js',
   '/js/wearables-oura.js',
   '/js/wearables-oura-auth.js',
   '/js/wearables-withings.js',
