+++ tsconfig.checkjs.json
@@ -389,7 +389,6 @@
     "js/wearables-runtime.js",
     "js/wearables-settings-runtime.js",
     "js/wearables-settings-groups.js",
-    "js/wearable-relay-consent.js",
     "js/wearables-settings-panel.js",
     "js/wearables-store.js",
     "js/wearables-strip-actions.js",
